doc_formatter_agent/requirements.txt
fastapi
uvicorn
python-docx
playwright
python-multipart
google-generativeai
requests
chromadb
sentence-transformers
language-tool-python
jinja2
aiofiles

